refactor(cli): extract installer build service
